[docker] Add feature to deploy a release identified by its name
The requirements.cfg file can have now a variable named RELEASE
which can contain the name of the release to be deployed. If found,
the versions of the components are downloaded from the directory
at https://github.com/Bitergia/mordred/docker/unified_releases.
If not found, it will deploy the versions specified in the
requirements.cfg file.