Improve the Exception handling for threads by using queues. Fix T2031
Previous implementation was not raising correctly the exceptions, due to
there was not method to control the exceptions launched from the
mother/father process. Current changes allow threads to share exceptions
using a queue, so the main process can wait and read if any exception
was raised and needs to be addressed.

The main process does not wait and check always, in this case the
exception is raised from the thread but the traceback is not printed,
as this is something done by the parent. This is something to be improved.