[config] Add variable number of configuration files
Now -c accepts a list of configuration files, instead of
only one. If only one is specified, it works as before.
If more than one is specified, all of them are read,
in order. The paramerters they have cascade, that is,
parameters set in later files overwrite those set in
previous files.