[task_collection][arthur] Measure and log the size of the dict which holds tha arthur raw items to be processed
In the current implementation, all the raw items for all data sources
in mordred are collected together in the same Redis queue. All those
items are read and stored in an in memory dict in mordred. The size
of this items dict could be huge and we need to manage it.