[task_identities] Don't load identities if autorefresh is pending
The autorefresh identities list is built from unify output in a SortingHat state.
If we change this status before the refresh is done, for example
breaking identities in a load with reset, the autorefresh list won't
be valid anymore. Thus, until there iare no autorfresh identities,
the SortingHat state won't be changed loading identities.