[task_identities] Execute unify just after the load of identities in the same task
To avoid enriching data with identities that are not yet unified,
after the loading of identities, execute a unify process.

It is critial when loading identities with reset, because after the
loading of identities, all the identities loaded from data sources
are unmerged and unaffiliated because of the reset.

If the refresh of identities is done in this status the identities
in the enriched index will be "reseted" so all affiliations and
merging of identities will be lost.