[tests] Replace mediawiki archive
This code replaces the archive to test mediawiki backend. Initially
the tests were executed against data generated from an older version
of mediawiki (< 1.27), which implied to mock the current datatime.
However, when performing integration tests the mocking strategy
couldn't be applied leading to failures in the tests.
Now, the data contained in the archive comes from a more recent
version of mediawiki (>= 1.27), which doesn't rely on the current
datetime.