Change name from mordred to sirmordred
I have (hopefully) changed all occurrences of mordred to sirmordred,
and Mordred to SirMordred (except when it clearly referred to
Mordred as a character, when I changed it to "Sir Mordred".

This includes the name of the Python package, of the Python module,
of the mordred,py file within the module, of the Mordred class,
and the driver program (bin/mordred).

Tests that used mordred have also been changed to sirmordred.