[config] Include sections for studies
This code allows to define sections for studies (onion,
demography and areas of code). In any backend section, a list
of studies can be declared. Every declared study must have its
parameters in a specific cfg section defined by
<type-of-the-study>:<study-id> (e.g., enrich_onion:1).
The id makes possible to execute the same types of studies for
a given backend.