[utils] Add micro-mordred
This code proposes a tiny version of mordred to execute just
once raw, merge identities, enrich, panels tasks of mordred
on a given set of backends (declared in a cfg file passed
as input). The micro-mordred overcomes the current limitations
of the p2o script in ELK, which is not able to execute
multiple studies for the input backend and requires constant
changes to align its logic with the ELK one. The micro-mordred
isn't supposed to require constant changes, since there is not
gap between its logic and the mordred one.

In order to execute the micro-mordred on git and github, an
example of the command is shown below:
`micro --raw --identities --enrich --panels
--cfg ./setup.cfg --backends git github:issue github:pull`.