[task_panels] Handle strict mode loading panels via setup.cfg
This code enhances `task_panels` to allow handling the way of
loading panels according to a parameter (`strict`) within the
setup.cfg. By default, the panels are loaded in strict mode, in
order to disable it, it suffices to set `strict = false` within
the section `panels` in the setup.cfg.