Change logging level for Perceval and GrimoireELK and disable their logging files
Change the log level of the Perceval and GrimoireELK libraries when the parameter
debug mode is set to False, this way the Mordred all is more readable. The files
with the logs for those libraries are no longer used, so the tool will just write
the logs to the current file all.log.