[task-enrich] Refactor `refresh_identities`
This code refactors the `refresh_identities` method to align it
with the recent changes in the sortinghat api. In a nutshell,
in order to reduce the memory footprint, the sortinghat api
returns a generator of identities (instead of lists),
the refresh_identities method manipulates this list to prepare
a filter used to update the data with the identities
information updated. Furthermore, the log message has been
modified to reduce the information printed.