[mordred] Remove global data sources param
This code removes the global data sources parameter, which is now
automatically assessed based on the projects.json. In a nutshell,
if a given data source appears in the section `unknown` of the
projects.json and in another section, that data source is considered
global. Thus, meaning that the all data is collected but just a portion
of it is finally enriched.