[config] Add param `retention_hours`
This code enhances the config with a new attribute: `retention_hours`, used
to enforce data retention policy. In a nutshell, the attribute allows to
declare the maximum number of hours after which the data must be removed
from the platform. The param is optional and is activated when setting
a positive int number.