[task_enrich] Pass the last timestamp of the index before enrichment
This code retrieves the metadata__timestamp value of the last item
inserted in the enriched index before looping over the repos which
data is stored in the same index. This is needed to make sure that
the incremental enrichment works for data sources that are
collected globally but only partially enriched.