[task_collection] Pass projects.json repo url to ELK
This code allows to forward the url of a specific repo within
the projects.json to elk (`feed_backend` function). This change
is needed to simplify the assignment of the repo to the project
defined in the projects.json. The function `feed_backend` now
accepts a new param (set to None by default to not break
existing functionalities) `projects_json_repo`.

Signed-off-by: Valerio Cosentino <valcos@bitergia.com>