[task_projects] Remove projects_last_diff
This code replaces the logic used to calculate the
diff between two projects.json files with a
comparison between the hashes of the files. This
change has been done since the previous comparison
took into account only the projects added/removed
from the files (and not the repos), and replicate
this at repo level would add some extra complexity
which is not needed.

Tests have been modified accordingly.

Signed-off-by: Valerio Cosentino <valcos@bitergia.com>