[task_identities] Support lazy load of SH organizations
This code enables the automatic reload of the organization file
(defined in the sortinghat section within the setup.cfg) when a
change is detected. Changes are detected as variations on the
hash derived from the file content, thus when a modification
occurs on the file, the hash changes and the file is
fully loaded again.

Signed-off-by: Valerio Cosentino <valcos@bitergia.com>