[task_identities] Support lazy load of SH identities files
This code enables the automatic reload of the identities files
(defined in the sortinghat section within the setup.cfg) when a
change is detected. Changes are detected as variations on the
hashes derived from the identities files content, thus when
a modification occurs on a given identities file, the hash changes
and the identities are loaded again.

Signed-off-by: Valerio Cosentino <valcos@bitergia.com>