Python healthcheck script added to utils folder
New script added to identify when the Mordred container should be restarted. The
script searches for a given error pattern in the log and when the error is found it
exits with a error code equal to 1. The first time the check is executed it does
not look backwards in time to avoid the container being restarted when it is not
needed. Once the healtcheck has failed, the following calls will exit with error
code 1 to make sure the status of the container is unhealthy until it is restarted.

A new dependency is added with the library file-read-backwards, useful to read the
log file from the most recent lines to the oldest ones. The file with the
requirements was changed to include this new library which license is MIT.

Signed-off-by: Luis Cañas-Díaz <lcanas@bitergia.com>