[githubql] Add new study "enrich_reference_analysis"
The purpose of this study is to gather all the issues and
pull requests which are mutually referenced. Once these
references are obtained, all of the events for the given
issue or pull request are updated with the corresponding
list of URLs from the referenced items.

README.md updated accordingly.
Tests added accordingly.

Signed-off-by: Miguel Ángel Fernández <mafesan@bitergia.com>