[mordred] Add flag/arg to filter which repos run
This commit adds an optional argument/flag to the collect/enrich tasks
and micro.py that can be specified with a list of repositories.  If set,
only repositories in the provided list will be processed during the
collect/enrich phase, enabling much quicker run times for large
projects.

This introduces only one breaking change in the micro.py file, in the
form of a new positional argument.  This was done in order to maintain
the style of arguments accepted by that function, and because it is much
less likely that micro.py will be used as a library.  All other methods
offer the new behavior as an optional kwarg which maintains the current
behavior if not specified.

For more discussion/motivation, see:
https://github.com/chaoss/grimoirelab-sirmordred/issues/511

Ammended 2021-08-05:  Add documentation to `get_raw` and `get_enrich`

Signed-off-by: Emi Simpson <emi@alchemi.dev>