[deps] Fix dependecies in requirements
After using pyproject.yml in grimoirelab-elk,
pip install requirements needs that url packages
have the same syntax with the same branch and not in
editable mode.
This commit fixes the dependencies that conflicts
with the new mode.
In future releases requirements.txt file will be
deprecated and poetry will be used so this is a
temporal fix.

Signed-off-by: Jose Javier Merchante <jjmerchante@bitergia.com>