[docker] Update docker base image to Python 3.8
This change is needed due to there are some problems
with GrimoireLab 0.3.0 and its required pandas version.
That version is not available in Python 3.9 so the
platform doesn't work and Python 3.9 is the default
version of Python in Debian Bullseye.

With this change, we set 'python:3.8-slim-bullseye'
as the base version of this image, to make sure
we use Python 3.8 as the default interpreter.

Signed-off-by: Santiago Dueñas <sduenas@bitergia.com>