[task_enrich] Allows to customize the name of the study aliases
This code allows you to customize the name of the study aliases
for the indexes. These names come from the aliases.json file now,
so the hardcoded names were removed.

i.e. In this aliases.json file the name of the study is
`enrich_onion` and the name of the alias is `all_onion`
```
"studies_aliases": {
    "enrich": [
      {
        "alias": "all_onion",
        "name": "enrich_onion"
      },
      ...
    ]
}
```

Test updated accordingly.

Signed-off-by: Quan Zhou <quan@bitergia.com>