Fix race condition on initial tasks
When mordred starts, it loads the dashboards to Kibana
and next, it reads the list of projects to analyze.
Due to a race condition, the thread that reads the list
of projects could not even start, so data backends won't
have any input to work with.

This commit fixes the problem adding isolated code that
will run the initial tasks. Once these tasks are finished,
the main loop will start the task manager.

Fixes chaoss/grimoirelab#585

Signed-off-by: Santiago Dueñas <sduenas@bitergia.com>