Fix identity refresh bug for all roles
This commit resolves an issue where role identities
(like `reviewed_by` or `merged_by` among others) were not
refreshed.
The bug occurred because the queries to OpenSearch filtered
items in which `author_uuid` was the individual pk, but not
when the `<role>_uuid` was the individual pk. That causes many
items that contain the `<role>_uuid` but not the `author_uuid`
not to be refreshed.

For example, the `Author` role was not affected because it
matched `author_uuid`, but roles like `reviewed_by` did not
match and were missed.

Signed-off-by: Jose Javier Merchante <jjmerchante@bitergia.com>